TarantoolOperations refactoring
