+++ src/main/java/org/springframework/data/tarantool/core/query/TarantoolRepositoryExecutor.java
@@ -1,5 +1,10 @@
 package org.springframework.data.tarantool.core.query;
 
+/**
+ * Helpers interface for executing repository queries
+ *
+ * @author Oleg Kuznetsov
+ */
 public interface TarantoolRepositoryExecutor {
 
     Object execute(final Object[] parameters);
