+++ src/main/java/org/springframework/data/tarantool/core/query/TarantoolRepositoryQueryObjectExecutor.java
@@ -2,6 +2,12 @@ package org.springframework.data.tarantool.core.query;
 
 import org.springframework.data.tarantool.core.TarantoolOperations;
 
+/**
+ * Helper for executing repository queries marked by {@link org.springframework.data.tarantool.repository.Query}
+ * annotation and having result as object or list of objects
+ *
+ * @author Oleg Kuzentsov
+ */
 public class TarantoolRepositoryQueryObjectExecutor implements TarantoolRepositoryExecutor {
 
     private final TarantoolOperations operations;
