+++ src/test/java/org/springframework/data/tarantool/entities/Address.java
@@ -3,12 +3,14 @@ package org.springframework.data.tarantool.entities;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import org.springframework.data.tarantool.core.mapping.Tuple;
 
 /**
  * @author Alexey Kuzin
  */
 @Data
 @Builder
+@Tuple("test_space")
 @EqualsAndHashCode
 public class Address {
     private String city;
